% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3_bn_to_graphNEL.R
\name{bn_to_graphNEL}
\alias{bn_to_graphNEL}
\title{Convert a bn object to graphNEL object}
\usage{
bn_to_graphNEL(graph_bn)
}
\arguments{
\item{graph_bn}{a \code{bn} object of Bayesian network}
}
\value{
a \code{graphNEL} object
}
\description{
Convert a bn object to graphNEL object while removing isolated nodes
}
\author{
Han Yu
}
