% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R1_data_chest.R
\docType{data}
\name{chest}
\alias{chest}
\title{A simulated data from the Chest Clinic example}
\format{
The data set \code{chest} contains two objects: 
\describe{
 \item{\code{data}}{a \code{data.frame} object of 10000 observations and 8 discrete variables.}
 \item{\code{dag}}{a \code{graphNEL} object specifying the network structure.}
}
}
\usage{
data(chest)
}
\description{
Simulated data from the Chest Clinic example (also known as the Asia example) 
from Lauritzen and Spiegelhalter, 1988.
}
\references{
Lauritzen and Spiegelhalter (1988) Local Computations with Probabilities on Graphical Structures and 
their Application to Expert Systems (with Discussion). J. Roy. Stat. Soc. 50, p. 157-224. \cr
\cr
Dethlefsen, C., & Hojsgaard, S. (2005). A common platform for graphical models in R: The gRbase package. 
Journal of Statistical Software, 14(17), 1-12.
}
