% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_data_emission1000.R
\docType{data}
\name{emission1000}
\alias{emission1000}
\title{A simulated data from the Emission example}
\format{
The data set \code{emission1000} contains two objects: 
\describe{
 \item{\code{data}}{a \code{data.frame} object of 1000 observations and 3 discrete variables and 6 continuous variables.}
 \item{\code{dag}}{a \code{graphNEL} object specifying the network structure.}
}
}
\usage{
data(emission1000)
}
\description{
Simulated data from the Emission example (also known as the Waste Incinerator example)
}
\references{
Lauritzen, S. L., & Jensen, F. (2001). 
Stable local computation with conditional Gaussian distributions. Statistics and Computing, 11(2), 191-203.
}
