% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E6_GetValues.R
\name{GetValue}
\alias{GetValue}
\title{Possible values of a discrete variable}
\usage{
GetValue(tree, var, message = TRUE)
}
\arguments{
\item{tree}{a \code{\linkS4class{ClusterTree}} object}

\item{var}{the variables to be queried}

\item{message}{type of desired distribution}
}
\value{
a \code{vector} of the possible values of discrete variable. If the variable is continuous, 
the returned value will be \code{NULL}.
}
\description{
Obtain all the possible values of a discrete variable.
}
\examples{
data(toytree)
GetValue(toytree, "HDL")

}
\author{
Han Yu
}
