% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTree.R
\name{PrintTree}
\alias{PrintTree}
\title{Print the cluster tree}
\usage{
PrintTree(tree)
}
\arguments{
\item{tree}{a \code{clustertree} object}
}
\description{
Print a \code{clustertree} object in a readable format
}
\details{
Print a \code{clustertree} object in a readable format by outputing cluster members
and linear potentials assigned to each cluster.
}
\examples{
data(toytree)
PrintTree(toytree)

}
\author{
Han Yu
}
\references{
Cowell, R. G. (2005). Local propagation in conditional Gaussian Bayesian networks. 
Journal of Machine Learning Research, 6(Sep), 1517-1550.
}
\seealso{
\code{\link{clustertree}}
}

