% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A8_qtlnet_to_graphNEL.R
\name{qtlnet_to_graphNEL}
\alias{qtlnet_to_graphNEL}
\title{Convert qtlnet to graphNEL object}
\usage{
qtlnet_to_graphNEL(data)
}
\arguments{
\item{data}{a \code{qtlnet} object}
}
\value{
\item{\code{graphNEL}}{a \code{graphNEL} object.}
}
\description{
Extract network structure from qtlnet object and convert to graphNEL object
}
\details{
This function extracts network structure from qtlnet object and convert to graphNEL object.
The example data can be downloaded from <https://github.com/hyu-ub/BayesNetBP>.
}
\examples{

\dontrun{
load(liverqtl.rda)
qtlnet_to_graphNEL(liverqtl$qtlnet.fit)
}

}
\author{
Han Yu
}
