% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E7_Sampler.R
\name{Sampler}
\alias{Sampler}
\title{Sampling from the Bayesian network}
\usage{
Sampler(tree, n)
}
\arguments{
\item{tree}{a \code{\linkS4class{ClusterTree}} object}

\item{n}{a \code{integer} number of observations to generate}
}
\value{
a \code{dataframe} of generated data
}
\description{
Sampling from the joint distribution of all applicable nodes in the Bayesian network.
}
\examples{

data(toytree)
Sampler(tree = toytree, n = 10)

}
\references{
Cowell, R. G. (2005). Local propagation in conditional Gaussian Bayesian networks.
Journal of Machine Learning Research, 6(Sep), 1517-1550.
}
\author{
Han Yu
}
