% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3_data_toytree.R
\docType{data}
\name{toytree}
\alias{toytree}
\title{A synthetic toy dataset}
\format{The data set contains a propagated \code{clustertree} object \code{toytree}, 
which is ready for evidence absorption and making queries.}
\usage{
data(toytree)
}
\description{
This is a synthetic dataset with five discrete and four continuouos
variables. The local models of this dataset were randomly generated.
Specifically, the coefficients of linear models were generated from \emph{N(0,1)},
the variances of the error terms from \emph{chi-square(1)}, and the conditional 
probabilities for discrete factors generated from \emph{uniform(0,1)} followed by 
normalization. The data set contains a propagated \code{clustertree} object, 
which is ready for evidence absorption and making queries.
}
