% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G5_PlotTree.R
\name{PlotTree}
\alias{PlotTree}
\title{Plot the cluster tree}
\usage{
PlotTree(tree, color = "gray90")
}
\arguments{
\item{tree}{a \code{\linkS4class{ClusterTree}} object}

\item{color}{nodes color}
}
\description{
Plot the structure of a \code{\linkS4class{ClusterTree}} object
}
\details{
Plot the structure of \code{clustertree} object, with the nodes labeled by corresponding
elimination node. The circles represent continuous clusters, while the boxes represent discrete clusters.
This function requires \code{Rgraphviz} package.
}
\examples{

\dontrun{
library("Rgraphviz")
data(toytree)
PlotTree(toytree)
}

}
\references{
Cowell, R. G. (2005). Local propagation in conditional Gaussian Bayesian networks.
Journal of Machine Learning Research, 6(Sep), 1517-1550. \cr
\cr
Yu H, Moharil J, Blair RH (2020). BayesNetBP: An R Package for Probabilistic Reasoning in Bayesian
Networks. Journal of Statistical Software, 94(3), 1-31. <doi:10.18637/jss.v094.i03>.

Cowell, R. G. (2005). Local propagation in conditional Gaussian Bayesian networks.
Journal of Machine Learning Research, 6(Sep), 1517-1550.
}
\author{
Han Yu
}
