% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3_data_toytree.R
\docType{data}
\name{toytree}
\alias{toytree}
\title{A ClusterTree Example of Liver Model}
\format{
The data set contains a propagated \code{\linkS4class{ClusterTree}} object \code{toytree}, 
which is ready for evidence absorption and making queries.
}
\usage{
data(toytree)
}
\description{
A propagated \code{\linkS4class{ClusterTree}} object named \code{toytree}, obtained from 
\code{\link{liver}} QTL data.
}
